Make  errors int

+++ c/src/base/ByteProducer.h
@@ -2,10 +2,11 @@
 #define C_BYTEPRODUCER_H
 
 #include "ByteListener.h"
+#include "VariableLengthArray.h"
 
 typedef struct ByteProducer {
   void (*next) (struct ByteProducer *self, Byte v);
-  void (*error) (struct ByteProducer *self, Byte e);
+  void (*error) (struct ByteProducer *self, int e);
   void (*complete) (struct ByteProducer *self);
   void (*add_listener) (struct ByteProducer *self, ByteListener *listener);
   void (*remove_listener) (struct ByteProducer *self, ByteListener *listener);
